requirements.txt
fastapi
uvicorn
python-dotenv
supabase
langdetect
groq
gtts
